feat: discord bot
